server/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
qrcode[pil]
python-multipart
python-dotenv
